+++ contracts/mocks/ExpMathMock.sol
@@ -2,29 +2,29 @@
 
 pragma solidity 0.7.6;
 
-import '../iq/ExpMath.sol';
+import "../iq/ExpMath.sol";
 
 contract ExpMathMock {
-    uint256 public result;
+    uint112 public result;
     uint256 public gas;
 
     function halfLife(
-        uint256 t0,
-        uint256 c0,
-        uint256 t12,
-        uint256 t
-    ) public pure returns (uint256) {
+        uint32 t0,
+        uint112 c0,
+        uint32 t12,
+        uint32 t
+    ) public pure returns (uint112) {
         return ExpMath.halfLife(t0, c0, t12, t);
     }
 
     function measure(
-        uint256 t0,
-        uint256 c0,
-        uint256 t12,
-        uint256 t
+        uint32 t0,
+        uint112 c0,
+        uint32 t12,
+        uint32 t
     ) public {
         uint256 tgas = gasleft();
-        uint256 tmp = halfLife(t0, c0, t12, t);
+        uint112 tmp = halfLife(t0, c0, t12, t);
         gas = tgas - gasleft();
         result = tmp;
     }
